+++ deeplearning4j-scaleout/spark/dl4j-spark-nlp/src/main/java/org/deeplearning4j/spark/models/embeddings/glove/Glove.java
@@ -26,9 +26,9 @@ import org.apache.spark.api.java.JavaSparkContext;
 import org.apache.spark.api.java.function.Function;
 import org.apache.spark.api.java.function.PairFunction;
 import org.apache.spark.broadcast.Broadcast;
-import org.deeplearning4j.berkeley.CounterMap;
-import org.deeplearning4j.berkeley.Pair;
-import org.deeplearning4j.berkeley.Triple;
+import edu.berkeley.nlp.util.CounterMap;
+import edu.berkeley.nlp.util.Pair;
+import edu.berkeley.nlp.util.Triple;
 import org.deeplearning4j.models.glove.GloveWeightLookupTable;
 import org.deeplearning4j.models.word2vec.VocabWord;
 import org.deeplearning4j.models.word2vec.wordstore.VocabCache;
@@ -177,7 +177,7 @@ public class Glove implements Serializable {
                                 (float) gloveWeightLookupTable.getMaxCount());
             }
             counts.add(new Triple<>(next.getFirst(), next.getSecond(),
-                            coOccurrenceCounts.getCount(next.getFirst(), next.getSecond())));
+                    (float)coOccurrenceCounts.getCount(next.getFirst(), next.getSecond())));
 
         }
